Add "--disable-makeinfo" option
This is helpful for those who have makeinfo 5.x installed.
This new version of makeinfo enforces more checks and thus causes not
the most recent binutils to fail on documentation building.

Also this option implicitly sets "--no-pdf" option to inhibit generation
of documentation in .pdf format because sources were not compiled in
documents on previous step.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>

Cc: Anton Kolesov <akolesov@synopsys.com>
Cc: Vineet Gupta <vgupta@synopsys.com>
Cc: Mischa Jonker <mjonker@synopsys.com>
Cc: Francois Bedard <fbedard@synopsys.com>